code health: centralize maskKey and improve backend architecture
- Extracted maskKey and string utilities to Backend/utils/strings.js
- Centralized OTP, cookie, and AI parsing utilities
- Refactored Socket.io into a dedicated service with atomic MongoDB operations
- Modularized CORS and CSRF configurations
- Implemented a centralized logger and replaced console calls in backend
- Added frontend cleanup for timers/Speech API and integrated ToastProvider
- Verified utility logic and ensured backward compatibility with frontend
