+++ Backend/controllers/project.controller.js
@@ -4,6 +4,7 @@ import userModel from '../models/user.model.js';
 import mongoose from 'mongoose';
 import { validationResult } from 'express-validator';
 import { logger } from '../utils/logger.js';
+import { PROJECT_CATEGORIES } from '../config/constants.js';
 
 export const getAllProject = async (req, res) => {
     try {
@@ -106,24 +107,6 @@ export const updateProjectSidebarSettings = async (req, res) => {
 // Get project counts by category
 export const getProjectCountsByCategory = async (req, res) => {
   try {
-    // List of all categories as in frontend
-    const allCategories = [
-      'DSA',
-      'Frontend Development',
-      'Backend Development',
-      'Fullstack Development',
-      'Code Review & Optimization',
-      'Testing & QA',
-      'API Development',
-      'Database Engineering',
-      'Software Architecture',
-      'Version Control & Git',
-      'Agile Project Management',
-      'CI/CD Automation',
-      'Debugging & Troubleshooting',
-      'Documentation Generation',
-      'Code Refactoring'
-    ];
     // Get logged-in user
     const loggedInUser = await userModel.findOne({ email: req.user.email }).lean();
     if (!loggedInUser) {
@@ -141,17 +124,17 @@ export const getProjectCountsByCategory = async (req, res) => {
     ]);
     // Convert to { [category]: count }
     const result = {};
-    allCategories.forEach(cat => {
+    PROJECT_CATEGORIES.forEach(cat => {
       result[cat] = 0;
     });
     counts.forEach(item => {
-      if (result.hasOwnProperty(item._id)) {
+      if (Object.prototype.hasOwnProperty.call(result, item._id)) {
         result[item._id] = item.count;
       }
     });
     res.status(200).json(result);
   } catch (err) {
-        console.error('getProjectCountsByCategory error:', err);
+        logger.error('getProjectCountsByCategory error:', err);
         res.status(500).json({ error: 'Internal server error' });
   }
 }
