+++ frontend/src/config/constants.js
@@ -1,0 +1,17 @@
+export const PROJECT_CATEGORIES = [
+    { title: 'DSA', description: 'Data Structures & Algorithms', icon: 'ri-bar-chart-fill' },
+    { title: 'Frontend Development', description: 'UI coding & client-side logic', icon: 'ri-code-box-line' },
+    { title: 'Backend Development', description: 'Server-side logic & integration', icon: 'ri-server-line' },
+    { title: 'Fullstack Development', description: 'End-to-end solutions', icon: 'ri-stack-line' },
+    { title: 'Code Review & Optimization', description: 'Improve and refactor code', icon: 'ri-git-pull-request-line' },
+    { title: 'Testing & QA', description: 'Automated tests & quality checks', icon: 'ri-test-tube-line' },
+    { title: 'API Development', description: 'Design and integrate APIs', icon: 'ri-terminal-box-line' },
+    { title: 'Database Engineering', description: 'Design and manage databases', icon: 'ri-database-2-line' },
+    { title: 'Software Architecture', description: 'Design scalable systems', icon: 'ri-building-3-line' },
+    { title: 'Version Control & Git', description: 'Source control best practices', icon: 'ri-git-branch-line' },
+    { title: 'Agile Project Management', description: 'Planning & task management', icon: 'ri-calendar-line' },
+    { title: 'CI/CD Automation', description: 'Build and deploy pipelines', icon: 'ri-settings-2-line' },
+    { title: 'Debugging & Troubleshooting', description: 'Identify and fix issues', icon: 'ri-bug-line' },
+    { title: 'Documentation Generation', description: 'Auto-generate docs & comments', icon: 'ri-book-2-line' },
+    { title: 'Code Refactoring', description: 'Improve structure without changing behavior', icon: 'ri-scissors-cut-line' }
+];
