+++ frontend/src/config/translations.js
@@ -1,0 +1,110 @@
+export const PROJECT_TRANSLATIONS = {
+  'en-US': {
+    addUsers: 'Add Users',
+    collaborators: 'Collaborators',
+    options: 'Options',
+    run: 'Run',
+    noFileSelected: 'No file selected.',
+    noCode: 'No code to display.',
+    settings: 'Settings',
+    language: 'Language',
+    aiAssistant: 'AI Assistant',
+    replyTo: 'Replying to',
+    send: 'Send',
+    searchMessages: 'Search messages...',
+    selectUser: 'Select User',
+    addCollaborators: 'Add Collaborators',
+    previewOptions: 'Preview Options',
+    editorSettings: 'Editor Settings (see main settings for more)',
+  },
+  'hi-IN': {
+    addUsers: 'यूज़र जोड़ें',
+    collaborators: 'सहयोगी',
+    options: 'विकल्प',
+    run: 'चलाएँ',
+    noFileSelected: 'कोई फ़ाइल चयनित नहीं है।',
+    noCode: 'कोड उपलब्ध नहीं है।',
+    settings: 'सेटिंग्स',
+    language: 'भाषा',
+    aiAssistant: 'एआई सहायक',
+    replyTo: 'को उत्तर दे रहे हैं',
+    send: 'भेजें',
+    searchMessages: 'संदेश खोजें...',
+    selectUser: 'यूज़र चुनें',
+    addCollaborators: 'सहयोगी जोड़ें',
+    previewOptions: 'पूर्वावलोकन विकल्प',
+    editorSettings: 'संपादक सेटिंग्स (अधिक के लिए मुख्य सेटिंग्स देखें)',
+  },
+  'es-ES': {
+    addUsers: 'Agregar usuarios',
+    collaborators: 'Colaboradores',
+    options: 'Opciones',
+    run: 'Ejecutar',
+    noFileSelected: 'Ningún archivo seleccionado.',
+    noCode: 'No hay código para mostrar.',
+    settings: 'Configuración',
+    language: 'Idioma',
+    aiAssistant: 'Asistente de IA',
+    replyTo: 'Respondiendo a',
+    send: 'Enviar',
+    searchMessages: 'Buscar mensajes...',
+    selectUser: 'Seleccionar usuario',
+    addCollaborators: 'Agregar colaboradores',
+    previewOptions: 'Opciones de vista previa',
+    editorSettings: 'Configuración del editor (ver configuración principal para más)',
+  },
+  'fr-FR': {
+    addUsers: 'Ajouter des utilisateurs',
+    collaborators: 'Collaborateurs',
+    options: 'Options',
+    run: 'Exécuter',
+    noFileSelected: 'Aucun fichier sélectionné.',
+    noCode: 'Aucun code à afficher.',
+    settings: 'Paramètres',
+    language: 'Langue',
+    aiAssistant: 'Assistant IA',
+    replyTo: 'En réponse à',
+    send: 'Envoyer',
+    searchMessages: 'Rechercher des messages...',
+    selectUser: 'Sélectionner un utilisateur',
+    addCollaborators: 'Ajouter des collaborateurs',
+    previewOptions: 'Options d’aperçu',
+    editorSettings: 'Paramètres de l’éditeur (voir les paramètres principaux pour plus)',
+  },
+  'de-DE': {
+    addUsers: 'Benutzer hinzufügen',
+    collaborators: 'Mitarbeiter',
+    options: 'Optionen',
+    run: 'Ausführen',
+    noFileSelected: 'Keine Datei ausgewählt.',
+    noCode: 'Kein Code zum Anzeigen.',
+    settings: 'Einstellungen',
+    language: 'Sprache',
+    aiAssistant: 'KI-Assistent',
+    replyTo: 'Antwort an',
+    send: 'Senden',
+    searchMessages: 'Nachrichten suchen...',
+    selectUser: 'Benutzer auswählen',
+    addCollaborators: 'Mitarbeiter hinzufügen',
+    previewOptions: 'Vorschauoptionen',
+    editorSettings: 'Editor-Einstellungen (siehe Haupteinstellungen für mehr)',
+  },
+  'ja-JP': {
+    addUsers: 'ユーザーを追加',
+    collaborators: '共同作業者',
+    options: 'オプション',
+    run: '実行',
+    noFileSelected: 'ファイルが選択されていません。',
+    noCode: '表示するコードがありません。',
+    settings: '設定',
+    language: '言語',
+    aiAssistant: 'AIアシスタント',
+    replyTo: '返信先',
+    send: '送信',
+    searchMessages: 'メッセージを検索...',
+    selectUser: 'ユーザーを選択',
+    addCollaborators: '共同作業者を追加',
+    previewOptions: 'プレビューオプション',
+    editorSettings: 'エディター設定（詳細はメイン設定を参照）',
+  },
+};
