+++ frontend/src/screens/Categories.jsx
@@ -3,6 +3,7 @@ import { motion, AnimatePresence } from 'framer-motion';
 import { useNavigate } from 'react-router-dom';
 import axios, { clearCsrfCache } from '../config/axios.js';
 import { ThemeContext } from '../context/theme.context.jsx';
+import { PROJECT_CATEGORIES } from '../config/constants.js';
 
 const Categories = () => {
   const navigate = useNavigate();
@@ -82,23 +83,7 @@ const Categories = () => {
   // Removed showTileSizeMenu, now handled by openDropdown
 
 
-  const categories = [
-    { title: 'DSA', description: 'Data Structures & Algorithms', icon: 'ri-bar-chart-fill' },
-    { title: 'Frontend Development', description: 'UI coding & client-side logic', icon: 'ri-code-box-line' },
-    { title: 'Backend Development', description: 'Server-side logic & integration', icon: 'ri-server-line' },
-    { title: 'Fullstack Development', description: 'End-to-end solutions', icon: 'ri-stack-line' },
-    { title: 'Code Review & Optimization', description: 'Improve and refactor code', icon: 'ri-git-pull-request-line' },
-    { title: 'Testing & QA', description: 'Automated tests & quality checks', icon: 'ri-test-tube-line' },
-    { title: 'API Development', description: 'Design and integrate APIs', icon: 'ri-terminal-box-line' },
-    { title: 'Database Engineering', description: 'Design and manage databases', icon: 'ri-database-2-line' },
-    { title: 'Software Architecture', description: 'Design scalable systems', icon: 'ri-building-3-line' },
-    { title: 'Version Control & Git', description: 'Source control best practices', icon: 'ri-git-branch-line' },
-    { title: 'Agile Project Management', description: 'Planning & task management', icon: 'ri-calendar-line' },
-    { title: 'CI/CD Automation', description: 'Build and deploy pipelines', icon: 'ri-settings-2-line' },
-    { title: 'Debugging & Troubleshooting', description: 'Identify and fix issues', icon: 'ri-bug-line' },
-    { title: 'Documentation Generation', description: 'Auto-generate docs & comments', icon: 'ri-book-2-line' },
-    { title: 'Code Refactoring', description: 'Improve structure without changing behavior', icon: 'ri-scissors-cut-line' }
-  ];
+  const categories = PROJECT_CATEGORIES;
 
 
   const [projectCounts, setProjectCounts] = useState({});
