+++ frontend/src/screens/Register.jsx
@@ -3,7 +3,6 @@ import ReCAPTCHA from 'react-google-recaptcha';
 import { Link, useNavigate } from 'react-router-dom';
 import { UserContext } from '../context/user.context';
 import { ThemeContext } from '../context/theme.context';
-import { useToast } from '../context/toast.context';
 import axios from '../config/axios';
 import anime from 'animejs';
 
@@ -19,7 +18,6 @@ const Register = () => {
     const [userId, setUserId] = useState('');
     const { setUser } = useContext(UserContext);
     const { isDarkMode } = useContext(ThemeContext);
-    const { showToast } = useToast();
     const navigate = useNavigate();
 
     const [errorMsg, setErrorMsg] = useState('');
@@ -164,12 +162,11 @@ const Register = () => {
             .then((res) => {
                 localStorage.setItem('token', res.data.token);
                 setUser(res.data.user);
-                showToast('Registration successful!', 'success');
                 setShowOtpModal(false);
                 navigate('/categories', { replace: true });
             })
             .catch(() => {
-                showToast('Invalid OTP. Please check your email and try again.', 'error');
+                alert('Invalid OTP. Please check your email and try again.');
             });
     }
